fix for type error on 64 bit systems
git-svn-id: svn://localhost/ardour2/trunk@837 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/session.cc
@@ -1728,13 +1728,13 @@ Session::new_audio_track (int input_channels, int output_channels, TrackMode mod
 		} while (track_id < (UINT_MAX-1));
 
 		if (input_auto_connect & AutoConnectPhysical) {
-			nphysical_in = min (n_physical_inputs, physinputs.size());
+			nphysical_in = min (n_physical_inputs, (uint32_t) physinputs.size());
 		} else {
 			nphysical_in = 0;
 		}
 		
 		if (output_auto_connect & AutoConnectPhysical) {
-			nphysical_out = min (n_physical_outputs, physinputs.size());
+			nphysical_out = min (n_physical_outputs, (uint32_t) physinputs.size());
 		} else {
 			nphysical_out = 0;
 		}
